requirements.txt
Flask==2.0.2
Flask-MySQL==1.5.2
